+++ internal/command/android.go
@@ -14,7 +14,7 @@ const (
 	// androidOS is the android OS name
 	androidOS = "android"
 	// androidImage is the fyne-cross image for the Android OS
-	androidImage = "fyneio/fyne-cross:1.1-android"
+	androidImage = "docker.io/fyneio/fyne-cross:1.2-android"
 )
 
 var (
@@ -71,13 +71,18 @@ func (cmd *Android) Run() error {
 
 	for _, ctx := range cmd.Context {
 
+		err := bumpFyneAppBuild(ctx)
+		if err != nil {
+			log.Infof("[i] FyneApp.toml: unable to bump the build number. Error: %s", err)
+		}
+
 		log.Infof("[i] Target: %s/%s", ctx.OS, ctx.Architecture)
 		log.Debugf("%#v", ctx)
 
 		//
 		// pull image, if requested
 		//
-		err := pullImage(ctx)
+		err = pullImage(ctx)
 		if err != nil {
 			return err
 		}