+++ internal/command/docker.go
@@ -2,6 +2,7 @@ package command
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 	"os"
 	"os/user"
@@ -12,6 +13,7 @@ import (
 	"github.com/fyne-io/fyne-cross/internal/icon"
 	"github.com/fyne-io/fyne-cross/internal/log"
 	"github.com/fyne-io/fyne-cross/internal/volume"
+
 	"golang.org/x/sys/execabs"
 )
 
@@ -25,8 +27,12 @@ const (
 // CheckRequirements checks if the docker binary is in PATH
 func CheckRequirements() error {
 	_, err := execabs.LookPath("docker")
+	if err == nil {
+		return nil
+	}
+	_, err = execabs.LookPath("podman")
 	if err != nil {
-		return fmt.Errorf("missed requirement: docker binary not found in PATH")
+		return fmt.Errorf("missed requirement: docker or podman binary not found in PATH")
 	}
 	return nil
 }
@@ -39,8 +45,48 @@ type Options struct {
 	Debug        bool     // Debug if true enable log verbosity
 }
 
+// engine returns the default engine name, if no engine is found it returns an empty string and the error.
+func engine() (string, error) {
+	if isEngineDocker() {
+		return "docker", nil
+	}
+
+	if isEnginePodman() {
+		return "podman", nil
+	}
+
+	return "", errors.New("no container engine found")
+}
+
+// isEnginePodman returns true if the engine is "podman"
+func isEnginePodman() bool {
+	_, err := execabs.LookPath("podman")
+	return err == nil
+
+}
+
+// isEngineDocker return true is the engine is "docker". If "docker" is an alias to podman, so it returns false.
+func isEngineDocker() bool {
+	execPath, err := execabs.LookPath("docker")
+	if err != nil {
+		return false
+	}
+	// if "docker" comes from an alias (i.e. "podman-docker") should not contain the "docker" string
+	out, err := execabs.Command(execPath, "--version").Output()
+	if err != nil {
+		return false
+	}
+	return strings.Contains(strings.ToLower(string(out)), "docker")
+}
+
 // Cmd returns a command to run in a new container for the specified image
 func Cmd(image string, vol volume.Volume, opts Options, cmdArgs []string) *execabs.Cmd {
+
+	// detect engine binary
+	engineBinary, err := engine()
+	if err != nil {
+		log.Fatal(err.Error())
+	}
 
 	// define workdir
 	w := vol.WorkDirContainer()
@@ -51,12 +97,30 @@ func Cmd(image string, vol volume.Volume, opts Options, cmdArgs []string) *execa
 	args := []string{
 		"run", "--rm", "-t",
 		"-w", w, // set workdir
-		"-v", fmt.Sprintf("%s:%s", vol.WorkDirHost(), vol.WorkDirContainer()), // mount the working dir
+		"-v", fmt.Sprintf("%s:%s:z", vol.WorkDirHost(), vol.WorkDirContainer()), // mount the working dir
 	}
 
 	// mount the cache dir if cache is enabled
 	if opts.CacheEnabled {
-		args = append(args, "-v", fmt.Sprintf("%s:%s", vol.CacheDirHost(), vol.CacheDirContainer()))
+		args = append(args, "-v", fmt.Sprintf("%s:%s:z", vol.CacheDirHost(), vol.CacheDirContainer()))
+	}
+
+	// handle settings related to engine
+	if isEnginePodman() {
+		args = append(args, "--userns", "keep-id", "-e", "use_podman=1")
+	} else {
+		// docker: pass current user id to handle mount permissions on linux and MacOS
+		if runtime.GOOS != "windows" {
+			u, err := user.Current()
+			if err == nil {
+				args = append(args, "-u", fmt.Sprintf("%s:%s", u.Uid, u.Gid))
+				args = append(args, "--entrypoint", "fixuid")
+				if !opts.Debug {
+					// silent fixuid if not debug mode
+					cmdArgs = append([]string{"-q"}, cmdArgs...)
+				}
+			}
+		}
 	}
 
 	// add default env variables
@@ -70,23 +134,13 @@ func Cmd(image string, vol volume.Volume, opts Options, cmdArgs []string) *execa
 		args = append(args, "-e", e)
 	}
 
-	// attempt to set fyne user id as current user id to handle mount permissions
-	// on linux and MacOS
-	if runtime.GOOS != "windows" {
-		u, err := user.Current()
-		if err == nil {
-			args = append(args, "-e", fmt.Sprintf("fyne_uid=%s", u.Uid))
-		}
-	}
-
 	// specify the image to use
 	args = append(args, image)
 
 	// add the command to execute
 	args = append(args, cmdArgs...)
 
-	// run the command inside the container
-	cmd := execabs.Command("docker", args...)
+	cmd := execabs.Command(engineBinary, args...)
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 
@@ -124,21 +178,75 @@ func goModInit(ctx Context) error {
 	return nil
 }
 
+// findGoFlags return the index of context where GOFLAGS is set, or -1 if it is not present.
+func findGoFlags(ctx Context) int {
+	for i, env := range ctx.Env {
+		if strings.HasPrefix(env, "GOFLAGS") {
+			return i
+		}
+	}
+	return -1
+}
+
+// findCgoLdFlags return the index of context where CGO_LDFLAGS is set, or -1 if it is not present.
+func findCgoLdFlags(ctx Context) int {
+	for i, env := range ctx.Env {
+		if strings.HasPrefix(env, "CGO_LDFLAGS") {
+			return i
+		}
+	}
+	return -1
+}
+
 // goBuild run the go build command in the container
 func goBuild(ctx Context) error {
 	log.Infof("[i] Building binary...")
 	// add go build command
-	args := []string{"go", "build"}
+	args := []string{"go", "build", "-trimpath"}
 
-	ldflags := ctx.LdFlags
 	// Strip debug information
 	if ctx.StripDebug {
-		ldflags = append(ldflags, "-w", "-s")
+		// ensure that CGO_LDFLAGS is not overwritten as they can be passed
+		// by the --env argument
+		if idx := findCgoLdFlags(ctx); idx > -1 {
+			// append the ldflags after the existing CGO_LDFLAGS
+			ctx.Env[idx] += " -w -s"
+		} else {
+			// create the CGO_LDFLAGS environment variable and add the strip debug flags
+			ctx.Env = append(ctx.Env, "CGO_LDFLAGS=-w -s")
+		}
 	}
 
+	ldflags := ctx.LdFlags
 	// add ldflags to command, if any
 	if len(ldflags) > 0 {
-		args = append(args, "-ldflags", fmt.Sprintf("'%s'", strings.Join(ldflags, " ")))
+		flags := make([]string, len(ldflags))
+		for i, flag := range ldflags {
+			parts := strings.Split(flag, "-") // because there could be several -X flags
+			for _, p := range parts {
+				if strings.HasPrefix(p, "X") {
+					// We must rebuild cases
+					// if the user pass "-ldflags "-X A.B=C" so we need to set it to "-X=A.B=C"
+					// if the user pass "-ldflags "-X=A.B=C" so we should not change it
+					sp := strings.SplitN(p, " ", 1)
+					if len(sp) == 2 {
+						args = append(args, "-ldflags=-X="+sp[1])
+					} else {
+						args = append(args, "-ldflags=-"+p)
+					}
+				} else {
+					// others flags can be set to GOFLAGS
+					flags[i] = "-ldflags=-" + p
+				}
+			}
+		}
+
+		// ensure that GOFLAGS is not overwritten as they can be passed
+		// by the --env argument
+		if idx := findGoFlags(ctx); idx > -1 {
+			// append the ldflags after the existing GOFLAGS
+			ctx.Env[idx] += " " + strings.Join(flags, " ")
+		}
 	}
 
 	// add tags to command, if any
@@ -371,7 +479,11 @@ func pullImage(ctx Context) error {
 	buf := bytes.Buffer{}
 
 	// run the command inside the container
-	cmd := execabs.Command("docker", "pull", ctx.DockerImage)
+	runner, err := engine()
+	if err != nil {
+		log.Fatal(err.Error())
+	}
+	cmd := execabs.Command(runner, "pull", ctx.DockerImage)
 	cmd.Stdout = &buf
 	cmd.Stderr = &buf
 
@@ -379,7 +491,7 @@ func pullImage(ctx Context) error {
 		log.Debug(cmd)
 	}
 
-	err := cmd.Run()
+	err = cmd.Run()
 
 	if ctx.Debug {
 		log.Debug(buf.String())