+++ internal/command/linux.go
@@ -14,10 +14,10 @@ const (
 	// linuxOS it the linux OS name
 	linuxOS = "linux"
 	// linuxImage is the fyne-cross image for the Linux OS
-	linuxImageAmd64 = "fyneio/fyne-cross:1.1-base"
-	linuxImage386   = "fyneio/fyne-cross:1.1-linux-386"
-	linuxImageArm64 = "fyneio/fyne-cross:1.1-linux-arm64"
-	linuxImageArm   = "fyneio/fyne-cross:1.1-linux-arm"
+	linuxImageAmd64 = "docker.io/fyneio/fyne-cross:1.2-base"
+	linuxImage386   = "docker.io/fyneio/fyne-cross:1.2-linux-386"
+	linuxImageArm64 = "docker.io/fyneio/fyne-cross:1.2-linux-arm64"
+	linuxImageArm   = "docker.io/fyneio/fyne-cross:1.2-linux-arm"
 )
 
 var (
@@ -69,13 +69,18 @@ func (cmd *Linux) Run() error {
 
 	for _, ctx := range cmd.Context {
 
+		err := bumpFyneAppBuild(ctx)
+		if err != nil {
+			log.Infof("[i] FyneApp.toml: unable to bump the build number. Error: %s", err)
+		}
+
 		log.Infof("[i] Target: %s/%s", ctx.OS, ctx.Architecture)
 		log.Debugf("%#v", ctx)
 
 		//
 		// pull image, if requested
 		//
-		err := pullImage(ctx)
+		err = pullImage(ctx)
 		if err != nil {
 			return err
 		}