+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog - Fyne.io fyne-cross
 
+## 1.2.0 - 07 Mar 2022
+
+### Added
+
+- Add support for FyneApp.toml #78
+- Add the ability to use podman #41
+- Update to use fixuid to handle mount permissions #42
+
 ## 1.1.3 - 02 Nov 2021
 
 ### Fixed